heart to Mouth

Released December 7, 2018

data("heart_to_mouth")

Format

The format is: List of 12 $ dreamcatcher : chr [1:31] "[Verse 1]" "Dreamcatcher, I can see you waving in the night alone" "Slow winter, and I feel I might get left out in the cold" "Ooh and I died today a little inside" ... $ when_i_m_over_you : chr [1:70] "[Verse 1]" "Ooh, ooh" "Wanna fight, wanna get right" "Wanna live, wanna never die" ... $ one_night_in_the_sun: chr [1:62] "[Verse 1]" "Get your one night in the sun" "Be sure to take your adulation" "Get your pinch just when the moment's right" ... $ girls_go_wild : chr [1:65] "[Verse 1]" "I've been caged, I've been hounded, I've been hunted and tamed" "I'm the outlaw of outside and ready to rage" "I'm in search of the heart I can eat to renew me, yeah" ... $ recovery : chr [1:47] "[Verse 1]" "Our room is dark, the blinds are shut tight" "And everything is still too much outside" "It may be over, but not tonight" ... $ the_power : chr [1:68] "[Intro]" "Woah" "Woah" "Yeah" ... $ dreamer : chr [1:76] "[Intro]" "Baby, you're a dreamer" "You don't know what I need, yeah" "Now, I know you got heart" ... $ house_on_fire : chr [1:47] "[Verse 1]" "I can't say yes or no" "We'll never know, just lay low" "The timing gets to me" ... $ hey_nice_to_know_ya : chr [1:67] "[Verse 1]" "It's been long, long drawn" "Got my heart broke in the car" "Parked outside the minibar" ... $ die_for_your_love : chr [1:70] "[Verse 1]" "Turn the lights up, nothin' to hide from" "Let's see those eyes" "Turn your glance up, girl, just man up" ... $ shaken : chr [1:70] "[Verse 1]" "All night, sippin' on a feelin'" "Won't lie, was nervous for this evenin'" "Then you arrive" ... $ special : chr [1:60] "[Verse 1]" "Yeah, the money and the fame" "Is all bullshit in the end" "Brings you nothing else but pain" ...

Source

Genius (https://genius.com/albums/Lp/Heart-to-mouth)

Examples

heart_to_mouth$dreamcatcher

#>  [1] "[Verse 1]"                                               
#>  [2] "Dreamcatcher, I can see you waving in the night alone"   
#>  [3] "Slow winter, and I feel I might get left out in the cold"
#>  [4] "Ooh and I died today a little inside"                    
#>  [5] "And I can see you waiting for your chance to go"         
#>  [6] ""                                                        
#>  [7] "[Verse 2]"                                               
#>  [8] "Dreamcatcher, when we're lyin' in our bed"               
#>  [9] "We're soarin', flyin'"                                   
#> [10] "Slow dancer, I been leaning on you"                      
#> [11] "Tryna' get some light"                                   
#> [12] "Ooh baby take away"                                      
#> [13] "This nightmare that stays awake"                         
#> [14] "And I fear we're in a maze of our design"                
#> [15] "(Oh, oh, oh, oh)"                                        
#> [16] ""                                                        
#> [17] "[Verse 3]"                                               
#> [18] "Dreamcatcher, I can see you with your eyes so far away"  
#> [19] "No answer from the mystic who took back our yesterday"   
#> [20] "A feather of destiny finding its way to me"              
#> [21] "Like a willow on the wind that cannot stay"              
#> [22] "That cannot stay"                                        
#> [23] "That cannot stay"                                        
#> [24] "(Oh, oh, oh, oh)"                                        
#> [25] "(Oh, oh, oh, oh)"                                        
#> [26] ""                                                        
#> [27] "[Outro]"                                                 
#> [28] "Dreamcatcher"                                            
#> [29] "Dreamcatcher"                                            
#> [30] "Dreamcatcher"                                            
#> [31] "Dreamcatcher"