Lost On You

Released May 5, 2017

data("lost_on_you")

Format

The format is: List of 10 $ muddy_waters : chr [1:53] "[Verse 1]" "We are kneeling at the rivers edge and tempting" "All the steps to follow closer right behind" "Is it only when you feel a part is empty" ... $ no_witness : chr [1:65] "[Verse 1]" "How do I get you outta my head?" "How do I get you back in my bed?" "Oh no, oh no" ... $ lost_on_you : chr [1:69] "[Verse 1]" "When you get older, plainer, saner" "Will you remember all the danger" "We came from?" ... $ up_against_me : chr [1:59] "[Verse 1]" "I wanted more" "It went away" "I asked the Lord" ... $ tightrope : chr [1:80] "[Verse 1]" "Hold on, I know you're scared" "But you're so close to heaven" "Eyes shut tight" ... $ other_people : chr [1:52] "[Verse 1]" "I never tried to be a hero" "You took us from the stars to zero" "" ... $ into_the_wild : chr [1:59] "[Verse 1]" "Are we on the lonely side?" "Say oh-oh-oh now" "The past long away" ... $ strange : chr [1:52] "[Verse 1]" "Hey" "Anyone can get love" "Anyone can get laid" ... $ death_valley : chr [1:54] "[Verse 1]" "Dime-store rock and roll" "Stretching out into forever" "I said, \"I don't know" ... $ you_want_it_all: chr [1:44] "[Verse 1]" "We been on this roundabout" "Swimming in these fields of doubt" "But there is no time now" ...

Source

Genius (https://genius.com/albums/Lp/Lost-on-you-international-version)

Examples

lost_on_you$muddy_waters

#>  [1] "[Verse 1]"                                                      
#>  [2] "We are kneeling at the rivers edge and tempting"                
#>  [3] "All the steps to follow closer right behind"                    
#>  [4] "Is it only when you feel a part is empty"                       
#>  [5] "That it’s gnawing at the corners of your mind"                  
#>  [6] ""                                                               
#>  [7] "[Pre-Chorus]"                                                   
#>  [8] "I will ask you for mercy"                                       
#>  [9] "I will come to you blind"                                       
#> [10] "What you’ll see is the worst me"                                
#> [11] "Not the last of my kind"                                        
#> [12] ""                                                               
#> [13] "[Chorus]"                                                       
#> [14] "In the muddy water we’re falling"                               
#> [15] "In the muddy water we’re crawling"                              
#> [16] "Hold me down"                                                   
#> [17] "Hold me now"                                                    
#> [18] "Sold me out"                                                    
#> [19] "In the muddy waters we’re falling"                              
#> [20] ""                                                               
#> [21] "[Verse 2]"                                                      
#> [22] "It is not clear why we choose the fire pathway"                 
#> [23] "Where we end is not the way that we had planned"                
#> [24] "All the spirits gather 'round like its our last day"            
#> [25] "To get across you know we’ll have to raise the sand"            
#> [26] ""                                                               
#> [27] "[Pre-Chorus]"                                                   
#> [28] "I will ask you for mercy"                                       
#> [29] "I will come to you blind"                                       
#> [30] "What you’ll see is the worst me"                                
#> [31] "I'm not the last of my kind"                                    
#> [32] ""                                                               
#> [33] "[Chorus]"                                                       
#> [34] "In the muddy water we’re falling"                               
#> [35] "In the muddy water we’re crawling"                              
#> [36] "Holds me down"                                                  
#> [37] "Hold me now"                                                    
#> [38] "Sold me out"                                                    
#> [39] "In the muddy water we’re falling (falling, falling, falling...)"
#> [40] ""                                                               
#> [41] "[Bridge]"                                                       
#> [42] "Don’t fail me now"                                              
#> [43] "Put your arms around me and pull me out"                        
#> [44] "I know I’m found"                                               
#> [45] "With your arms around me, oh save me now"                       
#> [46] ""                                                               
#> [47] "[Chorus]"                                                       
#> [48] "In the muddy water we’re falling"                               
#> [49] "In the muddy water we’re crawling"                              
#> [50] "Holds me down"                                                  
#> [51] "Hold me now"                                                    
#> [52] "Sold me out"                                                    
#> [53] "In the muddy water we’re falling"